Eliminate non-CUDA-API compiler warnings across LibTorchSharp and NativeAcceleration
- Remove #pragma once from .cpp files (8 files)
- Replace deprecated torch::IntList with at::IntArrayRef
- Fix NULL used for non-pointer types (cufftHandle defaults/comparisons, float returns)
- Remove ~25 unused/unreferenced variables and dead code
- Fix unsigned arithmetic bugs in tmax/tmin clamping (Xray.cu, Backproject.cu)
- Suppress #20054-D for benign __shared__ struct initialization
- Fix sign conversion in Remap.cu sentinel assignment
- Remove Clang-only -Wno-unused-local-typedef from CMakeLists.txt

+++ NativeAcceleration/gtom/src/ImageManipulation/Xray.cu
@@ -99,12 +99,12 @@ namespace gtom
 		{
 			uint samples = 0;
 			tfloat localmean = (tfloat)0;
-			for (int z = tmax(0, blockIdx.z - region); z <= tmin(dims.z - 1, blockIdx.z + region); z++)
+			for (int z = tmax(0, (int)blockIdx.z - region); z <= tmin(dims.z - 1, (int)blockIdx.z + region); z++)
 			{
-				for (int y = tmax(0, blockIdx.y - region); y <= tmin(dims.y - 1, blockIdx.y + region); y++)
+				for (int y = tmax(0, (int)blockIdx.y - region); y <= tmin(dims.y - 1, (int)blockIdx.y + region); y++)
 				{
 					size_t offsety = (z * dims.y + y) * dims.x;
-					for (int x = tmax(0, idx - region); x <= tmin(dims.x - 1, idx + region); x++)
+					for (int x = tmax(0, (int)idx - region); x <= tmin(dims.x - 1, (int)idx + region); x++)
 					{
 						if (x == idx && y == blockIdx.y && z == blockIdx.z)
 							continue;
